requirements.txt
Django==1.11.10
django-extensions==1.9.9
djangorestframework==3.7.7
pytz==2018.3
six==1.11.0
typing==3.6.4
